backend/requirements.txt
# LifeEmbedding Backend API Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# CORS
python-multipart==0.0.6

# Google Cloud
google-cloud-bigquery==3.13.0
google-cloud-aiplatform==1.38.0

# Data Processing
numpy==1.24.3
scikit-learn==1.3.2
umap-learn==0.5.4

# Utilities
python-dotenv==1.0.0
